[Feat] 카드 등록, 카드 조회, 포인트 잔액 조회, 결제 및 내역 생성

+++ Ssook_BE/src/main/java/com/stillalive/Ssook_BE/pay/repository/CardRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface CardRepository extends JpaRepository<Card, Long> {
+public interface CardRepository extends JpaRepository<Card, Integer> {
     Optional<Card> findByCardToken(String cardToken);
 
     Optional<Card> findByChild(Child child);
